# store: reduce memory footprint for chunks queries