--- a/pkg/pool/pool.go
+++ b/pkg/pool/pool.go
@@ -112,12 +112,11 @@ func (p *BucketedBytes) Put(b *[]byte) {
 			continue
 		}
 		*b = (*b)[:0]
+		p.mtx.Lock()
+		defer p.mtx.Unlock()
 		p.buckets[i].Put(b)
 		break
 	}
-
-	p.mtx.Lock()
-	defer p.mtx.Unlock()
 
 	// We could assume here that our users will not make the slices larger
 	// but lets be on the safe side to avoid an underflow of p.usedTotal.