@krya-kryak 2021-03-16
```
2021-03-16T12:19:18.5949904Z --- FAIL: TestStoreGateway (35.07s)
2021-03-16T12:19:18.5950809Z     --- FAIL: TestStoreGateway/query_works (25.86s)
2021-03-16T12:19:18.5951573Z === CONT  TestReceive/replication_with_outage
<...>
2021-03-16T12:19:18.7179785Z FAIL
2021-03-16T12:19:18.7180515Z FAIL	github.com/thanos-io/thanos/test/e2e	150.873s
2021-03-16T12:19:18.7181531Z ?   	github.com/thanos-io/thanos/test/e2e/e2ethanos	[no test files]
2021-03-16T12:19:18.7182110Z FAIL
2021-03-16T12:19:18.7183653Z ##[error]make: *** [Makefile:243: test-e2e] Error 1
2021-03-16T12:19:18.7192266Z ##[error]Process completed with exit code 2.
```
Failing e2e test `TestStoreGateway` seems to work locally (output cleaned up):
```
GOROOT=/usr/local/Cellar/go/1.16.2/libexec #gosetup
GOPATH=/Users/vkononov/go #gosetup
/usr/local/Cellar/go/1.16.2/libexec/bin/go test -c -o /private/var/folders/31/qr0fcfm56vs91g1fl524n63m0000gp/T/___TestStoreGateway_in_github_com_thanos_io_thanos_test_e2e github.com/thanos-io/thanos/test/e2e #gosetup
/usr/local/Cellar/go/1.16.2/libexec/bin/go tool test2json -t /private/var/folders/31/qr0fcfm56vs91g1fl524n63m0000gp/T/___TestStoreGateway_in_github_com_thanos_io_thanos_test_e2e -test.v -test.run ^\QTestStoreGateway\E$
=== RUN   TestStoreGateway
=== PAUSE TestStoreGateway
=== CONT  TestStoreGateway
--- PASS: TestStoreGateway (21.98s)
=== RUN   TestStoreGateway/query_works
    --- PASS: TestStoreGateway/query_works (5.05s)
=== RUN   TestStoreGateway/remove_meta.json_from_id1_block
    --- PASS: TestStoreGateway/remove_meta.json_from_id1_block (1.19s)
=== RUN   TestStoreGateway/upload_block_id5,_similar_to_id1
    --- PASS: TestStoreGateway/upload_block_id5,_similar_to_id1 (2.82s)
=== RUN   TestStoreGateway/delete_whole_id2_block_#yolo
    --- PASS: TestStoreGateway/delete_whole_id2_block_#yolo (3.03s)
PASS

Process finished with exit code 0
```