@GiedriusS 2021-03-29
> > LGTM, all comments have been addressed, is this still a WIP @krya-kryak? I would like to merge this nice optimization. Plus, I couldn't reproduce e2e test failures anymore so they are either flaky or GH Actions runners had some issues
> 
> @GiedriusS, thank you. I've set WIP with thoughts of optimizing CPU usage ( [#3937 (comment)](https://github.com/thanos-io/thanos/pull/3937#discussion_r596654131) ). Unfortunately, I was not able to dedicate some time to it this week :(
> If you consider current state of things good enough - I'll be happy to merge it and add performance optimizations (if any) later.

Given that most, if not all, CPU usage happen at the side of Thanos Query, I think that merging this as-is should be OK. But, let's wait for the opinion of others.